:bug #1958 【企业微信】获取应用详情接口返回的部门ID列表字段修改为Long类型

+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/bean/WxCpAgent.java
@@ -95,7 +95,7 @@ public class WxCpAgent implements Serializable {
   @Data
   public class Parties {
     @SerializedName("partyid")
-    private List<Integer> partyIds = null;
+    private List<Long> partyIds = null;
   }
 
   @Data
